feat: 完善 add to multi 逻辑

+++ YMHTTP/Classes/core/YMTimeoutSource.h
@@ -12,11 +12,13 @@ NS_ASSUME_NONNULL_BEGIN
 @interface YMTimeoutSource : NSObject
 
 @property (nonatomic, strong) dispatch_source_t rawSource;
-@property (nonatomic, assign) int milliseconds;
+@property (nonatomic, assign) NSInteger milliseconds;
 @property (nonatomic, strong) dispatch_queue_t queue;
 @property (nonatomic, strong) dispatch_block_t handler;
 
-- (instancetype)initWithQueue:(dispatch_queue_t)queue milliseconds:(int)milliseconds handler:(dispatch_block_t)handler;
+- (instancetype)initWithQueue:(dispatch_queue_t)queue
+                 milliseconds:(NSInteger)milliseconds
+                      handler:(dispatch_block_t)handler;
 
 @end
